Updating progress on the JS conversion
